新規テストファイルのパッケージ名をui配下の小文字命名規則に合わせる
masterのpackageリネーム(closes #110)後にマージしたため、RoutineCreateViewModelTest.kt
とRoutineEditViewModelTest.ktが旧パス/キャメルケースパッケージ名のまま取り込まれていたのを修正
